Replaced MultiMessage with the same class from MicroLib
Deleted the MultiMessage class.

Changed the MicroLib version to '37df70d6' in pom.xml to have access to the MultiMessage class.

Replaced all imports of the original MultiMessage class to the one in MicroLib. Also fixed some of the MicroLib imports to use the new file path.

Updated miscStartupProcedures() in PhantomWorlds.java to throw OutdatedServerVersionException because getLatestVersion() from MicroLib's UpdateChecker class throws an exception. In turn, I placed the call to miscStartupProcedures() in a try-catch block to catch the exception. I wasn't sure how you'd want this handled so I just left it be for now.